Try regularizing adjoint solve by hand
The dual form of SSA is degenerate around zero thickness, velocity, or
strain rate, so the linearization of the problem will be rank-deficient.
This means that the adjoint problem might not be solvable. Here I tried
regularizing it with some small multiple of the regularized
linearization `J_r`. Other things to try:
* use the LSQR solver
* Add a linear viscous component to the friction at low stresses